--- src/app/activities/id/page.tsx
@@ -1,5 +1,0 @@
-const Activitie = () => {
-  return <div>Activitie</div>;
-};
-
-export default Activitie;
