+++ src/app/activities/page.tsx
@@ -1,5 +1,18 @@
-const Activitie = () => {
-  return <div>Activitie</div>;
+'use client';
+
+import { useSearchParams } from 'next/navigation';
+
+const Activities = () => {
+  const searchParams = useSearchParams();
+  const region = searchParams.get('region');
+  const startTime = searchParams.get('startTime');
+  const keyword = searchParams.get('keyword');
+
+  return (
+    <div>
+      result::: region: {region || '-'} + startTime: {startTime || '-'} | keyword: {keyword || '-'}
+    </div>
+  );
 };
 
-export default Activitie;
+export default Activities;
