--- src/app/activity/[id]/page.tsx
@@ -1,6 +1,0 @@
-'use client';
-
-const ActiviyDetail = () => {
-  return <>Activity</>;
-};
-export default ActiviyDetail;
