--- src/app/activity/page.tsx
@@ -1,6 +1,0 @@
-'use client';
-
-const Activiy = () => {
-  return <>Activity</>;
-};
-export default Activiy;
